Avoid integer sanitizer warnings in chain.o
Summary:
This is a backport of [[https://github.com/bitcoin/bitcoin/pull/24146 | core#24146]]

Depends on D12815

Test Plan:
With UBSAN
`ninja && ninja check check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12816

+++ src/chain.cpp
@@ -105,7 +105,7 @@ int64_t GetBlockProofEquivalentTime(const CBlockIndex &to,
     if (r.bits() > 63) {
         return sign * std::numeric_limits<int64_t>::max();
     }
-    return sign * r.GetLow64();
+    return sign * int64_t(r.GetLow64());
 }
 
 /**
